#8459 폼 양식 / 신청서 / 처리할문서 데이터 구조 통합 : 파일 업로드쪽 존재하는 경우에만 저장하도록 if 구문 추가함

+++ src/main/kotlin/co/brainz/framework/fileTransaction/service/AliceFileService.kt
@@ -168,15 +168,16 @@ class AliceFileService(
                 val fileLocEntity = aliceFileLocRepository.getOne(fileDataId[index].toLong())
                 val filePath = Paths.get(fileLocEntity.uploadedLocation + File.separator + fileLocEntity.randomName)
                 val tempPath = getDir("temp", fileLocEntity.randomName)
-                Files.move(tempPath, filePath, StandardCopyOption.REPLACE_EXISTING)
-                fileLocEntity.uploaded = true
-
-                try {
-                    aliceFileLocRepository.save(fileLocEntity)
-                } catch (e: Exception) {
-                    logger.error("{}", e.message)
-                    Files.move(filePath, tempPath, StandardCopyOption.REPLACE_EXISTING)
-                    throw AliceException(AliceErrorConstants.ERR, e.message)
+                if (Files.exists(tempPath)) {
+                    Files.move(tempPath, filePath, StandardCopyOption.REPLACE_EXISTING)
+                    fileLocEntity.uploaded = true
+                    try {
+                        aliceFileLocRepository.save(fileLocEntity)
+                    } catch (e: Exception) {
+                        logger.error("{}", e.message)
+                        Files.move(filePath, tempPath, StandardCopyOption.REPLACE_EXISTING)
+                        throw AliceException(AliceErrorConstants.ERR, e.message)
+                    }
                 }
             }
         }
